Fragments implemented Clickable topBar temporarily disabled
