+++ .idea/misc.xml
@@ -13,6 +13,8 @@
         <entry key="..\:/Users/Radmira/AndroidStudioProjects/YAMF/app/src/main/res/drawable/menu_baclground.xml" value="0.22135416666666666" />
         <entry key="..\:/Users/Radmira/AndroidStudioProjects/YAMF/app/src/main/res/layout/activity_details.xml" value="0.36666666666666664" />
         <entry key="..\:/Users/Radmira/AndroidStudioProjects/YAMF/app/src/main/res/layout/activity_main.xml" value="0.33" />
+        <entry key="..\:/Users/Radmira/AndroidStudioProjects/YAMF/app/src/main/res/layout/fragment_details.xml" value="0.36666666666666664" />
+        <entry key="..\:/Users/Radmira/AndroidStudioProjects/YAMF/app/src/main/res/layout/fragment_home.xml" value="0.33" />
         <entry key="..\:/Users/Radmira/AndroidStudioProjects/YAMF/app/src/main/res/layout/movie_item.xml" value="0.36666666666666664" />
         <entry key="..\:/Users/Radmira/AndroidStudioProjects/YAMF/app/src/main/res/menu/nav_menu.xml" value="0.36666666666666664" />
         <entry key="..\:/Users/Radmira/AndroidStudioProjects/YAMF/app/src/main/res/menu/toolbar.xml" value="0.36666666666666664" />
