+++ app/src/main/AndroidManifest.xml
@@ -10,9 +10,6 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.YAMF">
         <activity
-            android:name=".DetailsActivity"
-            android:exported="false" />
-        <activity
             android:name=".MainActivity"
             android:exported="true">
             <intent-filter>
