+++ app/src/main/res/layout/fragment_home.xml
@@ -1,0 +1,115 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    tools:context=".HomeFragment">
+
+
+
+    <androidx.cardview.widget.CardView
+        android:id="@+id/topBar"
+        android:layout_width="match_parent"
+        android:layout_height="@dimen/topBarHeight"
+        android:layout_margin="@dimen/topBarMargin"
+        app:cardCornerRadius="@dimen/topBarCornerRadius"
+        app:cardElevation="@dimen/topBarElevation"
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintEnd_toEndOf="parent">
+
+        <LinearLayout
+            android:layout_width="match_parent"
+            android:layout_height="100dp"
+            android:orientation="horizontal">
+
+
+            <ImageButton
+                android:id="@+id/image1"
+                android:layout_width="@dimen/cardview_default_elevation"
+                android:layout_height="wrap_content"
+                android:layout_weight="1"
+                android:background="@drawable/arcane"
+                android:clickable="true"
+                android:stateListAnimator="@animator/poster_selector"
+                android:contentDescription="@string/imageDescription"
+                android:scaleType="centerCrop"
+                android:focusable="true" />
+
+            <ImageButton
+                android:id="@+id/image2"
+                android:layout_width="@dimen/cardview_default_elevation"
+                android:layout_height="wrap_content"
+                android:layout_weight="1"
+                android:background="@drawable/bladerunner2049"
+                android:clickable="true"
+                android:stateListAnimator="@animator/poster_selector"
+                android:contentDescription="@string/imageDescription"
+                android:scaleType="centerCrop"
+                android:focusable="true" />
+
+            <ImageButton
+                android:id="@+id/image3"
+                android:layout_width="@dimen/cardview_default_elevation"
+                android:layout_height="wrap_content"
+                android:layout_weight="1"
+                android:background="@drawable/swmando"
+                android:clickable="true"
+                android:stateListAnimator="@animator/poster_selector"
+                android:contentDescription="@string/imageDescription"
+                android:scaleType="centerCrop"
+                android:focusable="true" />
+
+            <ImageButton
+                android:id="@+id/image4"
+                android:layout_width="@dimen/cardview_default_elevation"
+                android:layout_height="wrap_content"
+                android:layout_weight="1"
+                android:background="@drawable/indy"
+                android:clickable="true"
+                android:stateListAnimator="@animator/poster_selector"
+                android:contentDescription="@string/imageDescription"
+                android:scaleType="centerCrop"
+                android:focusable="true" />
+
+            <ImageButton
+                android:id="@+id/image5"
+                android:layout_width="@dimen/cardview_default_elevation"
+                android:layout_height="wrap_content"
+                android:layout_weight="1"
+                android:clickable="true"
+                android:stateListAnimator="@animator/poster_selector"
+                android:background="@drawable/spiderverse"
+                android:contentDescription="@string/imageDescription"
+                android:scaleType="centerCrop"
+                android:focusable="true" />
+        </LinearLayout>
+    </androidx.cardview.widget.CardView>
+
+    <TextView
+        android:id="@+id/disclaimer"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="@dimen/disclaimerMargin"
+        android:layout_marginBottom="@dimen/disclaimerMargin"
+        android:autoLink="web"
+        android:linksClickable="true"
+        android:text="@string/disclaimer"
+        android:textSize="16sp"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@id/topBar" />
+
+    <androidx.recyclerview.widget.RecyclerView
+        android:id="@+id/main_recycler"
+        android:layout_width="0dp"
+        android:layout_height="0dp"
+        android:layout_margin="16dp"
+        app:layout_constraintTop_toBottomOf="@id/disclaimer"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        tools:listitem="@layout/movie_item"
+        />
+</androidx.constraintlayout.widget.ConstraintLayout>
