+++ app/src/main/res/values/strings.xml
@@ -12,4 +12,6 @@
     <string name="menu_favorites_title">Favorites</string>
     <string name="menu_watch_later_title">Watch later</string>
     <string name="menu_collections_title">Collections</string>
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
